requirements.txt
# Core ML / Fine-tuning
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.3.3
transformers==4.57.0
accelerate==1.10.1
datasets==4.1.1
trl==0.23.1
sentence-transformers==5.1.1
umap-learn==0.5.9.post2
hdbscan==0.8.40
torch==2.5.1
bertopic==0.17.3

# LLM Integration
google-generativeai==0.8.5
openai==2.1.0
anthropic==0.69.0
pydantic==2.11.9

# Data Cleaning / NLP
regex==2025.9.18
nltk==3.9.2
spacy==3.7.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Visualization & Utilities
tqdm==4.67.1
matplotlib==3.10.6
seaborn==0.13.2

# Environment & Database
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
